tests/wpt/tests/element-timing/pseudo-background-image.html

<!DOCTYPE HTML>
<head>
<meta charset=utf-8>
<title>Element Timing: observe element with a background image in a pseudo element</title>
<style>
body {
  margin: 0;
}
#target::before {
  content: '';
  width: 100px;
  height: 100px;
  background: url(resources/square100.png);
  display: inline-block;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/element-timing-helpers.js"></script>
<script>
  let beforeRender = performance.now();
  async_test(function (t) {
    assert_implements(window.PerformanceElementTiming, "PerformanceElementTiming is not implemented");
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        const pathname = window.location.origin + '/element-timing/resources/square100.png';
        checkElement(entry, pathname, 'my_div', 'target', beforeRender,
            document.getElementById('target'));
        checkRect(entry, [0, 100, 0, 100]);
        checkNaturalSize(entry, 100, 100);
      })
    );
    observer.observe({entryTypes: ['element']});
  }, 'Element with elementtiming attribute and a ::before background image is observable.');
</script>
</head>
<body>
<div id='target' elementtiming='my_div'></div>
</body>
